Amended LOAD_DEMAND_DATASETS to include 01/01/2004
The reason the first window used to start on Jan 2nd (and the total number of windows being 352) was that Jan 1st didn't have a 00:00 data point (which was later "interpolated", along with Feb 29th) so the entire day was discarded. 

This issue was fixed back in 3341612211a6974681ec2bc32685ffe013374e96, but it was not reflected in the subsequent code.

Also, removed the unnecessary rewinding of `chp_notAveraged`.