Some fixes related to more compact data being loaded from files.
Previously we would compute the demand estimates for all building types, and would only use the one corresponding to the requested building type. Since we now only compute the demands only for the requested building to begin with, there's no need for the 2nd index which corresponds to the building type (`iB`) in many places.

Also, some internal variable renames.