Bugfix: incorrect index when accessing true demands
So far, the moving-average demand was accessed at the same day index as the the true demand. The issue with this can be demonstrated using an example: suppose today is the 15, then past demands from 01-14 are averaged (this window ends at midnight between the 14th and the 15th); when the true demands are accessed, we need to take directly the day of the 15th (which ends at midnight between the 15th and 16th). Thus the correct index for the true demands is +1.